Stop braking if the X+Y accelerations are low enough before the timeout
