+++ src/main/java/org/teamresistance/frc/Robot.java
@@ -6,7 +6,7 @@ import org.strongback.command.Command;
 import org.strongback.command.CommandGroup;
 import org.strongback.components.ui.FlightStick;
 import org.strongback.hardware.Hardware;
-import org.teamresistance.frc.command.HaltDrivingCommand;
+import org.teamresistance.frc.command.BrakeCommand;
 import org.teamresistance.frc.command.HoldAngleCommand;
 import org.teamresistance.frc.command.StrafeCommand;
 import org.teamresistance.frc.subsystem.drive.Drive;
@@ -49,7 +49,7 @@ public class Robot extends IterativeRobot {
     // Drive straight, pause for 2s, then strafe 90 degrees
     reactor.onTriggeredSubmit(leftJoystick.getButton(7), () -> CommandGroup.runSequentially(
         new StrafeCommand(drive, 0, 1.5),
-        new HaltDrivingCommand(drive, 1.5),
+        new BrakeCommand(drive, IO.gyro, 1.5),
         new StrafeCommand(drive, 90, 1.5)
     ));
 
@@ -57,7 +57,7 @@ public class Robot extends IterativeRobot {
     reactor.onTriggered(leftJoystick.getButton(3), () -> Strongback.submit(Command.cancel(drive)));
 
     // Lock the drive motors and hopefully stop the robot
-    reactor.onTriggeredSubmit(leftJoystick.getButton(5), () -> new HaltDrivingCommand(drive, 0.5));
+    reactor.onTriggeredSubmit(leftJoystick.getButton(5), () -> new BrakeCommand(drive, IO.gyro, 0.5));
   }
 
   @Override
